requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1
djangorestframework==3.11.1
future==0.18.2
idna==2.10
Markdown==3.2.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
urllib3==1.25.10
